+++ ui/src/core/components/Radio/RadioButtons/__tests__/RadioButtons.spec.tsx
@@ -16,7 +16,7 @@
 
 import React from 'react';
 import { render, screen } from 'unit-test/testUtils';
-import RadioButtons from '../';
+import RadioButtons from '..';
 import userEvent from '@testing-library/user-event';
 
 const props = {
